Rework how built-in types are loaded in the compiler
In TopDownAnalyzerFacadeForJVM, we now always use the "load built-ins
from module dependencies" behavior that was previously only enabled with
a dedicated CLI argument. However, sometimes we compile code without
kotlin-stdlib in the classpath, and we don't want everything to crash
because some standard type like kotlin.Unit hasn't been found.

To mitigate this, we add another module at the end of the dependencies
list, namely a "fallback built-ins" module. This module loads all
built-in declarations from the compiler's class loader. This prevents
the compiler from crashing, but compiling the code against built-ins
found in the compiler is still discouraged, so we report an error if
anything is resolved to a declaration from this module.

Also introduce a new CLI argument -Xsuppress-missing-builtins-error
specifically to suppress this error and to allow compiling code against
compiler's own built-ins.