Remove obsolete check for safecast in elvis
Previously, we had special treatment for cases like
'x as? Foo ?: return' - while analyzing elvis, we explicitly checked if
left-hand side is safecast, and if right was a jumpout, we add smartcast
info manually.

Since 1.2, we have "SafeCastCheckBoundSmartCasts" feature, which is more
general form of this logic and already covers such cases, so it is not
necessary anymore to hoop around safecasts in elvis specifically.