[NI] Refactor test data for contracts
Currently, we have following logic for if-analysis: if type of one
branch is ErrorType, then take data flow info of the other one as the
resulting data flow info at the if-exit.

This means that *we do get smartcast* for the following example:

fun foo(x: Int?, b: Boolean) {
  if (b) x.unresolved() else x!!
  x.inc() // smartcasted!
}

Contracts test data had similar code, where last expression was
meant to test if effect system-provided flow is not leaking outside of
if-branches.

However, because of old inference quirks, "unresolved" branch actually
had a type, so flows from the both branches were correctly merged
(obviously, "unresolved" branch had EMPTY flow).

In NI, "unresolved" branch has error type (which is intended behaviour
currently), so we get "bogus" leakage of flow. To prevent this, all such
functions were split into two: one checks that there are no incorrect
smartcast in a wrong branch, and the other one checks that if only correct
branch is present, then flow is not leaking.