[NI] Turn off KnownTypeParameterSubstitutor for NI
In OI, for special calls (!!, elvis, if, when) we've created synthethic
function with type-parameter T. Then, we've forced equality T =
EXPECTED_TYPE (if any), to tuck in diagnostic reporting from whole
expression to particular arguments.

In fact, this is kinda hacky. Moreover, this synthethic equality
constraint can harm new inference. So, we turn this off in NI.

This has two implications in NI:

- For some tests, range of reported TYPE_MISMATCH becomes wider
- For some tests, smartcasts behaviour has been changed (because
expected types has been changed)
Also there is some one-case error with "inaccessibleType"