Specialize iterator-based loop over ranges in bytecode
This is required for inline high-order functions in Kotlin stdlib to
work effectively with ranges. E.g., in code such as

  (1..5).forEach { ... }

the loop is inside inline function 'forEach'. Instead of using explicit
counter, this loop would allocate an iterator and invoke virtual methods
for 'hasNext' and 'next'. HotSpot is "good enough" at scalarizing local
objects and inlining virtual methods, however, performance cost for
computationally intensive code is ~(10-15)%, see KT-18802. This cost is
even more evident on Android, where JIT is significantly more
resource-constrained.

So we detect an iterator-based loop on an IntRange and LongRange in the
bytecode (after '(1..5).forEach { ... }' is compiled, with 'forEach'
inlined), and transform it to the counter-based loop if possible.

Note that this transformation is disabled if a function uses an
explicitly created iterator (such as 'Iterable<T>.reduce').

 #KT-18802 Fixed